+++ CHANGELOG.md
@@ -7,11 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
 - `Collection`: Pass `worldwide` property to `spatial-extents` slot.
 
 ### Fixed
 
 - `SearchableList`:  Minor CSS fix for badges
+- Minor CSS fixes for overflowing content, especially in collections
 
 ## [2.4.1] - 2020-08-03