+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.5.0] - 2020-08-13
+
 ### Added
 
 - `Collection`: Pass `worldwide` property to `spatial-extents` slot.
@@ -421,7 +423,8 @@ First stable release.
 Please see the [Releases](https://github.com/Open-EO/openeo-vue-components/releases) for changelogs prior to v1.0.0.
 
 
-[Unreleased]: https://github.com/Open-EO/openeo-vue-components/compare/v2.4.1...HEAD
+[Unreleased]: https://github.com/Open-EO/openeo-vue-components/compare/v2.5.0...HEAD
+[2.5.0]: https://github.com/Open-EO/openeo-vue-components/compare/v2.4.1...v2.5.0
 [2.4.1]: https://github.com/Open-EO/openeo-vue-components/compare/v2.4.0...v2.4.1
 [2.4.0]: https://github.com/Open-EO/openeo-vue-components/compare/v2.3.4...v2.4.0
 [2.3.4]: https://github.com/Open-EO/openeo-vue-components/compare/v2.3.3...v2.3.4